pyproject.toml
[tool.poetry]
name = "genjax-localization-tutorial"
version = "0.1.0"
description = ""
authors = ["Matthew Huebert <mhuebert@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
jupytext = "^1.16.1"
genjax = {version = "0.7.0.post4.dev0+eacb241e"  , source = "gcp" }
# genstudio = {version = "2024.11.021", source = "gcp"}
genstudio = {path = "../genstudio", develop = true}
ipykernel = "^6.29.3"
matplotlib = "^3.8.3"
anywidget = "^0.9.7"
watchfiles = "^0.21.0"
markdown = "^3.6"

[[tool.poetry.source]]
name = "gcp"
url = "https://us-west1-python.pkg.dev/probcomp-caliban/probcomp/simple/"
priority = "explicit"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["genjax-localization-tutorial"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
